requirements.txt
flask==2.3.3
tensorflow==2.15.0
numpy==1.26.4
pillow==10.0.1
gunicorn==21.2.0
